pyproject.toml
[project]
name = "ckan-pycsw"
version = "0.1.0"
description = "Populate PyCSW using data from a CKAN catalog"
authors = [
    {name = "mjanez", email = "mnl.janez@gmail.com"},
]
dependencies = [
    "pygeometa>=0.13.1",
    # pycsw 3.0 installed from GitHub (not yet available on PyPI)
    "pycsw @ git+https://github.com/geopython/pycsw.git@master",
    # pycsw 3.0 standalone requirements (Flask-based multi-API server)
    "flask>=3.1.2",
    "pygeofilter>=0.2.0",
    "pygeoif>=1.0.0",
    "setuptools>=65.6.0",
    "sqlalchemy<2.0.0",
    "requests>=2.28.1",
    "shapely>=1.8.5.post1",
    "pyyaml>=6.0",
    "ptvsd>=4.3.2",
    "debugpy>=1.6.6",
    "simplejson>=3.19.1",
    "apscheduler>=3.10.1",
    "psutil>=5.9.5",
    "urllib3>=1.26.20",
]
requires-python = ">=3.9"
license = {text = "GPLv3"}
[project.optional-dependencies]
prod = [
    "gunicorn>=20.1.0",
]

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
